BREAKING CHANGE: switched to makefile+pyproject+flit as build system
This removes the previous support for `conda` environments. On the other hand,
users don't have to install Anaconda anymore; it's sufficient to have `make`
and `python` installed, and everything else bootstraps itself as needed.
Moreover, this switch makes it easier to use newer versions of Python,
which these days come with signficant performance benefits.

On servers, it's still recommended to use IntelPython instead, and just
manually install the dependencies in `pyproject.toml` into that distribution.